mmap: Allow an existing GuestMemoryMmap to be extended
The public API of vm-memory is pretty limited regarding the actions
that can be performed on a GuestMemoryMmap. The only allowed action
is to create a new instance of a GuestMemoryMmap based on a list of
memory ranges, but there are use cases where we need to extend an
existing instance by appending more memory regions to the existing
ones.

This change creates a new public function that can be used from any
caller in order to add a new memory region to an existing guest memory.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>